Moved the LDAP commands into a single file.
