+++ unsync_ldap3/unsync_ldap3/__init__.py
@@ -1,4 +1,4 @@
 __version__ = '0.0.1'
 
 unsync_prefix = 'ldap'
-unsync_commands = ['connection.connect', 'import.ldap_import']
+unsync_commands = ['commands.connect', 'commands.ldap_import']
